Finish login setup
